+++ CHANGELOG.md
@@ -7,6 +7,11 @@ The file format of it is based on [Keep a Changelog](http://keepachangelog.com/e
 For public Changelog covering all changes done to Pipedrive’s API, webhooks and app extensions platforms, see [public Changelog](https://pipedrive.readme.io/docs/changelog) with discussion area in [Developers Community](https://devcommunity.pipedrive.com/c/documentation/changelog/19).
 
 ## [Unreleased]
+## Changed
+- Updated `POST /products`: parameter `name` is required
+- Updated `PUT /deals/:dealId/products`: parameter `quantity` and `item_price` are required
+- Updated `POST /deals/:dealId/products`: parameter `quantity` and `item_price` are required
+- 
 
 ## 13.2.5
 ## Changed